requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-apscheduler==0.6.2
django-filter==22.1
django-modeltranslation==0.18.9
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
